Removed ChessPiece.availablePieces
It was causing an issue by which a piece would ignore pieces that were moved after it, being able to go through them.

Since it's been removed, it was neccesary to change how Game checks if it's possible to move to a given space. Also, the non-static updateAvailableMoves also needed to be updated.