Changed GameBoard to not be static, and more
GameBoard is not static now. Instead, Game is static, and Game has a field for its GameBoard.

A bunch of properties were moved from GameBoard to Game, mainly those that have to do with the state of the game; those neccesary to generate a FEN string.

Most of the code to generate or interpret a FEN string has also been moved from GameBoard to Game.

Also substantially changed TakenPieces. It no longer keeps track of what pieces have been taken. Instead, GameBoard does that, and TakenPieces asks GameBoard for it.