Fixed an oversight
When detecting a Checkmate, if there is a check but no checkmate, Game.goodMoves should be updated to contain all Coordinates that would block the check. However, if the king is able to move, it exited before calculating that.
That's been fixed by checking if goodMoves is null after finding a check. If it is null, then we calculate the moves.