Solved a few problems around checks
The king was allowed to move to attacked spaces. This happened because when asking enemy pieces for their moves, they saw that a king was in check, and only returned moves that blocked that check.

It's been solved by adding a field checkedKing: WorB. Now pieces will check if the king in check is their own before trying to block the check.

Additionaly, solved an issue where if your king was in check, your pieces were allowed to move without blocking the check so long as they were capturing a piece. This happened because of a missing if statement.

IMPORTANT: updateAvailableMoves now returns all attacked cells., This means that it might return cells which hold allied pieces. In order to disallow taking these pieces, cullAvailableMoves must be called. It removes from availableMoves all those moves that would take a piece of the same color as the selected one.