implemented castling and found an issue
There MIGHT be an issue with undoing when you have loaded a board. It would need to be checked further.

It seems to be fixed. But beware nonetheless.

in order to "fix" that, another field was added, that tracks the move we start with. So we have currentMoveNum and startingMoveNum