Fixed a bug explained below.
When promoting, the promoted piece was selected. This was intended. However, in selecting the piece, all its available moves were calculated and shown. This was not intended.

This issue has been fixed by creating a new method in ChessMatch that works as a simple setter for matchState.selectedPiece

This, however, is somewhat dangerous, and this method should only be used when absolutely certain that it is the right one for the case, since generally changing this attribute demands changes in others as well (which setSelectedPiece manages, but this method doesn't)